Refactor: 트랙 페이지 컴포넌트 디렉토리 구조 변경
